Added working code for prototype and breadboard
Prototype is made to client expectations and to the best of my knowledge
should work to make the relays fire properly with the button timers.

Breadboard version is a bare-bones button timer with all the whistles
removed. It can be tested using a breadboard, a pull-up/pull-down
resistor and a jumper wire to 5v.